web/requirements.txt
appnope==0.1.0
decorator==4.0.9
Django==1.9.2
django-contrib-comments==1.6.2
django-model-utils==2.4
django-pagedown==0.1.1
django-redis-cache==1.6.5
djangorestframework==3.3.2
PyGithub
gnureadline==6.3.3
gunicorn==19.3.0
httplib2==0.9.2
ipdb==0.9.0
ipython==4.1.1
ipython-genutils==0.1.0
Markdown==2.6.5
mock==1.3.0
oauthlib==1.0.3
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6
ptyprocess==0.5.1
Pygments==2.1.1
python-dateutil==2.4.2
python-instagram==1.3.2
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.1.0
tweepy==3.5.0
